dht/afr: Modify itransform/deitransform to prevent loss of brick-id incase of both dht & afr involved.
With Change-Id:Ieba9a7071829d51860b7c131982f12e0136b9855, dht itansform/deitransform
was improved to encode 64-bit brick offset along with the brick-id in the global d_off.

More details regarding this change are in : http://review.gluster.org/#/c/4711/

But now with afr using the same itransform/deitransform logic, the brick-id stored in the
afr_global_d_off gets zeroed out when re-encoded in dht. This happens only when the offsets are huge
(i.e with ext4 filesystem) as in such cases, the low n bits are replaced with brick-id which in
turn gets replaced with afr_subvol_id when re-encoded in dht, where
       n = log2(N)
       N = no. of DHT/AFR subvolumes.

To avoid this, instead of using the last 'n' bits, we have decided to store the brick/subvol-id in
the first 'n' bits leaving the TOP_BIT i.e, the brick/subvol offset will be first shifted by 'n' bits
to the right and then the brick-id is copied into those first 'n' bits.

This approach as well may result in the loss of few lower bits of the brick_offset. But as mentioned
in the review#4711, since "both EXT4/XFS are tolerant in terms of the accuracy of the value presented
back in seekdir(). i.e, a seekdir(val) actually seeks to the entry which has the "closest" true offset.",
this approach too seems to work.

Also there is an issue with small offsets as well. The smaller offset in the AFR translator
can result in a HUGE offset when passed to DHT which may again result in the loss of brick_id.
To prevent this, added an extra check in afr_itransform to consider such values as well for HUGE d_off.

Here is the modified algorithm-

  #define BACKEND_D_OFF_BITS 63
  #define PRESENT_D_OFF_BITS 63

  #define SHIFT_BITS (max (0, (BACKEND_D_OFF_BITS - PRESENT_D_OFF_BITS + 1)))
  #define TOP_BIT (ONE << (PRESENT_D_OFF_BITS - 1)) // 0x4000000000000000

  N = no. of DHT/AFR subvolumes.
  n = log2(N)

ENCODING :-

          hi_mask = 0x7fffffffffffffff;
          hi_mask = ~(hi_mask >> (n + 1));

        DHT :-
          if (d_off_brick_or_subvol & hi_mask) {
                /* HUGE d_off */
                brick_or_subvol_id = (brick_or_subvol_id << (PRESENT_D_OFF_BITS - 1 - n));
                d_off_global = (TOP_BIT | (d_off_brick_or_subvol >> (SHIFT_BITS + n)) | brick_or_subvol_id);
          } else {
                /* SMALL d_off */
                d_off_global = ((d_off_brick_or_subvol * N) + brick_or_subvol_id);
          }

        AFR :-
          if ((d_off_brick_or_subvol * N) & hi_mask) {
                /* HUGE d_off */
                brick_or_subvol_id = (brick_or_subvol_id << (PRESENT_D_OFF_BITS - 1 - n));
                d_off_global = (TOP_BIT | (d_off_brick_or_subvol >> (SHIFT_BITS + n)) | brick_or_subvol_id);
          } else {
                /* SMALL d_off */
                d_off_global = ((d_off_brick_or_subvol * N) + brick_or_subvol_id);
          }
DECODING :-

           hi_mask = 0x7fffffffffffffff;
           off_mask = (hi_mask >> (1+ n));
           host_mask = ~(off_mask);

           if (d_off_global & TOP_BIT) {
                /* HUGE d_off */
                d_off_brick_or_subvol = ((d_off_global & off_mask) << (SHIFT_BITS + n));
                brick_or_subvol_id = (((d_off_global & ~TOP_BIT) & host_mask) >> (PRESENT_D_OFF_BITS - 1 - n));
           } else {
                /* SMALL d_off */
                brick_or_subvol_id = (d_off_global % N);
                d_off_brick_or_subvol  = (d_off_global / N);
           }

Change-Id: I01bbe10a2bd77e54e1de000ade7df42d23791d1d
Signed-off-by: Soumya Koduri <skoduri@redhat.com>