gfapi: Providing support for the applications to apply LEASE LOCKs
Defined new lock routine "glfs_common_lock" which can be used
by the applications to take lease locks/share reservations by using
the new lock flags introduced. Note that this routine can be used to
take byte-range locks as well.

gfapi then maps these flags to appropriate Gluster defined flags which
are then passed to the glusterfs server as part of the syncop_lk request.

procotol/server, on receiving the LEASE_LK flag, sets a dict 'set_lease'
in the xdata for the other xlators to differentiate the lease locks from
the byte-range locks.

Note - This patch is not yet final and still WIP.

Change-Id: I14eeb59a025a01cdbb5f16a6e8c979418e472a20
Signed-off-by: Soumya Koduri <skoduri@redhat.com>
Signed-off-by: Anand Subramanian <anands@redhat.com>