Upcall: New xlator defined to store various states and send cbk events
Framework on the server-side, to handle certain state of the files
accessed and send notifications to the clients connected.

A generic and extensible framework, used to maintain states in
the glusterfsd process for each of the files accessed
(including the clients info doing the fops) and send
notifications to the respective glusterfs clients incase of
any change in that state.

Few of the use-cases (currently supported) by this infrastructure are:
        * Inode Update/Invalidation
        * lease locks

In addition 'rpc' transport details are stored in 'client_t' details which
will be used by the server to send callback requests.

Feature page: http://www.gluster.org/community/documentation/index.php/Features/Upcall-infrastructure

Note - This patch is not yet final and still WIP.

Change-Id: Ie3d724be9a3419fcf18901a753e8ec2df2ac802f
Signed-off-by: Soumya Koduri <skoduri@redhat.com>